Various bug fixes
Fixed the !addquotes command. It can only be used by mods, and it
actually works now. Added a 25 second cooldown to the !quote command.
Updated the !xp command to show user level and difference needed to
level up. Updated !topfive and !myrank command to not include me in the
rankings